Refactor: Add tips and note fields to shift form
- Added `tips` and `note` fields to the shift input form.
- Updated `ShiftFormFragment` to include `EditText` fields for tips and notes, and to handle their input.
- Modified `ShiftFormViewModel` to include `tips` and `note` in `UiState` and `calculateShift` logic.
- Updated `ShiftInputModel` to include `tips` and `note` fields.
- Adjusted `ShiftMappers` to include `tips` and `taxRate` in `ShiftFinanceInputEntity` and `ShiftFinanceInput`.
- Changed `rubToCents` to `dollarsToCents` in `ShiftInputToDomain.kt` for currency conversion.
- Added string resources for "Tips" and "Note" in English and Russian.

+++ app/src/main/java/com/hsact/taxilog/ui/shift/ShiftInputModel.kt
@@ -1,7 +1,7 @@
 package com.hsact.taxilog.ui.shift
 
 data class ShiftInputModel(
-    var carId: Long = 0,
+    var carId: Int = 0,
     var carName: String = "",
     var date: String = "",
     var timeStart: String = "",
